requirements.txt
backports.zoneinfo
certifi
cffi
charset-normalizer
idna
pycparser
urllib3
asgiref
argon2-cffi
cryptography
defusedxml
Django
djangorestframework
djangorestframework_simplejwt
djongo
elasticsearch-dsl
mysqlclient
oauthlib
pandas
pdfplumber
PyJWT
PyPDF2~=2.0
pymongo
python-dotenv
python3-openid
pytz
qrcode
requests
requests-oauthlib
sqlparse
django-allauth
django-allauth-2fa
django-crispy-forms==1.13.0
django-embed-video
django-elasticsearch-dsl
django-extensions
django-otp
django-werkzeug
django-db-logger
django-csp
django-axes
django-widget-tweaks
django-debug-toolbar
faker
pymysql
pytesseract
pdf2image
waitress
django-redis
redis
rake-nltk
BERTopic
